refactor(mcp): rename connection sync history UI module
