+++ airbyte/mcp/interactive/_connection_sync_history_ui.py
